requirements.txt
argcomplete==1.11.1
black==19.10b0
colorlog==4.1.0
docopt==0.6.2
flake8==3.8.2
flake8-import-order==0.18.1
isort==4.3.21
Jinja2==2.11.2
mypy==0.780
pipdeptree==1.0.0
pytest==5.4.3
PyYAML==5.3.1
rich==2.2.4
virtualenv==20.0.21
yarg==0.1.9
